More bootstrapping, here work on the handler and slight changes in the state

+++ hangman-tp-py/code/state.py
@@ -3,6 +3,8 @@
 
 import hashlib
 
+from enum import Enum
+
 from cbor2 import dumps, loads
 
 from sawtooth_sdk.processor.exceptions import InternalError
@@ -16,25 +18,43 @@ def _make_hm_address(name):
         hashlib.sha512(name.encode("utf-8")).hexdigest()[:64]
 
 
+class GameState(Enum):
+    ongoing = 1
+    won = 2
+    lost = 3
+
+
 class Game:
-    def __init__(self, name="", word="", misses="", host="", guesser=""):
+    def __init__(self, name="", word="", misses="", hits="", host="", guesser="", state=GameState.ongoing):
         self.name = name
         self.word = word
         self.misses = misses
+        self.hits = hits
         self.host = host
         self.guesser = guesser
+        self.state = state
 
     @classmethod
     def from_dict(cls, d):
-        return cls(d["name"], d["word"], d["misses"], d["host"], d["guesser"])
+        return cls(
+            d["name"],
+            d["word"],
+            d["misses"],
+            d["hits"],
+            d["host"],
+            d["guesser"],
+            d["state"]
+        )
 
     def to_dict(self):
         return {
             "name": self.name,
             "word": self.word,
             "misses": self.misses,
+            "hits": self.hits,
             "host": self.host,
             "guesser": self.guesser,
+            "state": self.state,
         }
 
 
@@ -43,8 +63,12 @@ class HmState:
         self._context = context
 
     def delete_game(self, name):
-        address = _make_hm_address(name)
-        self._context.delete_state([address], timeout=TIMEOUT)
+        game = self.get_game(name)
+        if game:
+            address = _make_hm_address(name)
+            self._context.delete_state([address], timeout=TIMEOUT)
+        else:
+            raise KeyError
 
     def set_game(self, name, game):
         address = _make_hm_address(name)
@@ -64,4 +88,4 @@ class HmState:
             state_de = loads(state_s)
             return Game.from_dict(state_de[-1])
         else:
-            return Game()
+            return None
